PCMI research git/vscode_python_data_starter/requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.1
